transaction-gc,c

+++ Grocerry_Management/System/models.py
@@ -57,7 +57,7 @@ class Bill(models.Model):
 class Transaction(models.Model):
     bill = models.ForeignKey(Bill, on_delete=models.CASCADE, related_name='transactions')  # Added related_name here
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.IntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=2)
     amount = models.DecimalField(max_digits=10, decimal_places=2)
     timestamp = models.DateTimeField(auto_now_add=True)
